got a details view working

+++ src/components/AdminTableRow/AdminTableRow.jsx
@@ -8,9 +8,19 @@ import { Envelope } from 'react-bootstrap-icons';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
+import Carousel from 'react-bootstrap/Carousel';
 import axios from 'axios';
+import swal from 'sweetalert';
+import Photos from './Photos';
+import Modal from 'react-bootstrap/Modal';
+import Container from 'react-bootstrap/Container';
+import Col from 'react-bootstrap/Col';
+import Badge from 'react-bootstrap/Badge'
+import {check2Circle} from 'react-bootstrap-icons'
+import * as Icon from 'react-bootstrap-icons';
 
 function AdminTableRow({ order }) {
+
     const dispatch = useDispatch();
     const [name, setName] = useState('');
     const [email, setEmail] = useState('');
@@ -18,14 +28,31 @@ function AdminTableRow({ order }) {
     const [toggleName, setToggleName] = useState(false);
     const [toggleEmail, setToggleEmail] = useState(false);
 
+    const handleShowModal = () => {
+        setModal(!modal);
+    }
+
     const deleteOrder = () => {
-        axios.delete(`/api/order/delete/${order.id}`)
-            .then((response) => {
-                dispatch({ type: 'FETCH_ORDERS' })
-            })
-            .catch((error) => {
-                console.log(`HEY MITCH - COULDN'T DELETE THE ORDER: ${error}`);
-            })
+        swal({
+            title: "Are you sure?",
+            text: `You are about to delete ${order.name}'s order... u sho 'bout dat? `,
+            icon: "warning",
+            dangerMode: true,
+            button: "DELETE IT"
+        })
+            .then(willDelete => {
+                if (willDelete) {
+                    axios.delete(`/api/order/delete/${order.id}`)
+                        .then((response) => {
+                            dispatch({ type: 'FETCH_ORDERS' })
+                        })
+                        .catch((error) => {
+                            console.log(`HEY MITCH - COULDN'T DELETE THE ORDER: ${error}`);
+                        })
+                }
+                else return;
+            });
+
     }
 
     const editMode = () => {
@@ -55,8 +82,6 @@ function AdminTableRow({ order }) {
             editMode();
         }
     };
-
-    const store = useSelector((store) => store);
     const formatDate = (orderDate) => {
         const date = new Date(orderDate);
         const options = { month: "short", day: "numeric", year: "numeric" }
@@ -65,46 +90,87 @@ function AdminTableRow({ order }) {
     }
 
     return (
-        <tr>
-            <td>
-                {formatDate(order.order_date)}
-            </td>
-            <td>
-                {toggleName ? <Form.Control
-                    onKeyPress={handleKeypress}
-                    onChange={((e) => { setName(e.target.value) })}
-                    value={name}
-                ></Form.Control> : order.name}
-            </td>
-            <td className="text-center">
-                {toggleEmail ? <Form.Control
-                    onKeyPress={handleKeypress}
-                    onChange={((e) => { setEmail(e.target.value) })}
-                    value={email}
-                ></Form.Control> : order.email}
-            </td>
-            <td>
-                {order.complete ? "Sent (Resend)" : "Pending"}
-            </td>
-            <td>
-                <ButtonGroup>
-                    <Button
-                        onClick={editMode}
-                        variant={toggleEmail ? "dark" : "outline-dark"}
-                    >
-                        <PencilSquare fontSize="2rem" />
-                    </Button>
-                    <Button
-                        onClick={deleteOrder}
-                        variant="outline-dark">
-                        <Trash fontSize="2rem" />
-                    </Button>
-                    <Button variant="outline-dark">
-                        <ZoomIn fontSize="2rem" />
-                    </Button>
-                </ButtonGroup>
-            </td>
-        </tr>
+        <>
+            <Modal
+                size="lg"
+                show={modal}
+                onHide={() => setModal(!modal)}
+                backdrop="static"
+                id="modal"
+            >
+                <Modal.Header closeButton>
+                    <Modal.Title>
+                        <h1>Pictures From {order.name}'s Order</h1>
+                    </Modal.Title>
+                </Modal.Header>
+                <Modal.Body className="modalTitle">
+                    <Carousel>
+                        {order.array_agg.map((image, i) => {
+                            return (
+                                <Carousel.Item key={i}>
+                                    <Photos image={image} />
+                                </Carousel.Item>
+                            )
+                        })}
+                    </Carousel>
+                </Modal.Body>
+            </Modal>
+            <tr>
+                <td>
+                    {formatDate(order.order_date)}
+                </td>
+                <td>
+                    {toggleName ? <Form.Control
+                        onKeyPress={handleKeypress}
+                        onChange={((e) => { setName(e.target.value) })}
+                        value={name}
+                    ></Form.Control> : order.name}
+                </td>
+                <td className="text-center">
+                    {toggleEmail ? <Form.Control
+                        onKeyPress={handleKeypress}
+                        onChange={((e) => { setEmail(e.target.value) })}
+                        value={email}
+                    ></Form.Control> : order.email}
+                </td>
+                <td>
+                    {order.complete ? 
+                    <Badge pill variant="success">
+                        Sent    <Icon.Check2Circle />
+                    </Badge>
+                    : <Badge pill variant="warning">
+                    Pending <Envelope />
+                </Badge>}
+                </td>
+                <td>
+                    <ButtonGroup>
+                        <Button variant="outline-dark">
+                            <Envelope
+                            variant={order.complete ? "outline-dark" : "secondary"}
+                            fontSize="2rem" />
+                        </Button>
+                        <Button
+                            onClick={editMode}
+                            variant={toggleEmail ? "dark" : "outline-dark"}
+                        >
+                            <PencilSquare fontSize="2rem" />
+                        </Button>
+                        <Button variant={modal ? "dark" : "outline-dark"}>
+                            <ZoomIn
+                                onClick={() => {
+                                    handleShowModal()
+                                }}
+                                fontSize="2rem" />
+                        </Button>
+                        <Button
+                            onClick={deleteOrder}
+                            variant="outline-dark">
+                            <Trash fontSize="2rem" />
+                        </Button>
+                    </ButtonGroup>
+                </td>
+            </tr>
+        </>
     );
 }
